requirements.txt
# Core runtime

# Deep learning
torch
torchvision

# Hugging Face
transformers
diffusers
accelerate
safetensors

# UI
gradio

# Imaging
pillow

# Utilities
pyyaml
gitpython
numpy

# Optional (performance / diagnostics)
psutil